requirements.txt
# Step 2 — Wazuh Indexer client (OpenSearch SDK)
opensearch-py>=2.7.0
python-dotenv>=1.0.0

# Step 5 — agent loop
anthropic>=0.40.0
langgraph>=0.2.0

# Step 6 — ChromaDB knowledge store + structured logging
chromadb>=0.5.0
python-json-logger>=2.0.7
# ChromaDB 0.5.x bundles an opentelemetry build that only accepts protobuf <4
protobuf>=3.19,<4

# Tests
pytest>=8.0.0
